+++ Books App/BooksMobile/Book.d.ts
@@ -4,3 +4,5 @@ type Book = {
     authors: string[];
     isbn: string;
 };
+
+type BookProvider = "googleBooksSearch" | "openLibrarySearch"
