+++ Books App/BooksMobile/screens/SearchScreen/queries.ts
@@ -1,0 +1,33 @@
+import { gql } from "@apollo/client";
+
+export const searchQuery = gql`
+  query SearchBooks($q: String) {
+    googleBooksSearch(q: $q, country: "US") {
+      items {
+        id
+        volumeInfo {
+          authors
+          averageRating
+          description
+          imageLinks {
+            thumbnail
+          }
+          title
+          subtitle
+          industryIdentifiers {
+            identifier
+            type
+          }
+        }
+      }
+    }
+    openLibrarySearch(q: $q) {
+      docs {
+        author_name
+        title
+        cover_edition_key
+        isbn
+      }
+    }
+  }
+`;
